feat: New method find_by_user_card_number added, function preventig the addition of a card with the same number added

+++ backend/controllers/LoyaltyCardController.php
@@ -8,11 +8,6 @@ class LoyaltyCardController {
         //Récupere et décode les données json envoyées par l'utilisateur
         $data = json_decode(file_get_contents("php://input"), true);
         //Verifie si tous les champs obligatoires sont remplis
-        if (!isset($data["establishment"]) || empty($data["establishment"])) {
-            http_response_code(400);
-            echo json_encode(["error" => "Le nom de l'établissement est requis"]);
-            return;
-        }
         
         if(!isset($data["card_number"]) || empty($data["card_number"]) || empty($data["user_id"])) {
             http_response_code(400);
@@ -21,12 +16,19 @@ class LoyaltyCardController {
         }else{
             //Crée une instance du modèle et enregistre le numero de la carte dans la base de données
             $loyaltyCardModel = new LoyaltyCard();
+            $userId = $data["user_id"];
+            $cardNumber = $data["card_number"];
+            $existingCard = $loyaltyCardModel->find_card_by_user_and_number($userId, $cardNumber);
+            if($existingCard) {
+                http_response_code(400);
+                echo json_encode(["error" => "Ce numéro de carte existe déjà pour cet utlisateur"]);
+                return;
+            }
             $loyaltyCardModel->create_card([
                 "card_number" => $data["card_number"],
                 "user_id" => $data["user_id"],
                 "establishment" => $data["establishment"],
-                "delivery_date" => $data["delivery_date"],
-                "expiry_date" => $data["expiry_date"]
+                "expiry_date" => $data["expiry_date"],
             ]);
             echo json_encode(["success" => "La carte a été ajouté" ]);
             
@@ -40,7 +42,7 @@ class LoyaltyCardController {
         if($cards) {
             echo json_encode(["success" => true, "cards" => $cards]);
         }else {
-        echo json_encode(["error" => "Aucune carte de fidélité"]);
+            echo json_encode(["error" => "Aucune carte de fidélité"]);
         }
     }
